Implemented correct decimating (scipy's version did not work), by manually applying low-pass-filter and np.interp() for reducing sampling rate.
